Attempt rename of crate
